refactor(web): separate render handlers - switch to xml marshaler

+++ persistence/postgres.go
@@ -177,7 +177,7 @@ func (db *postgresDatabase) QueryTorrents(
 	epoch int64,
 	orderBy OrderingCriteria,
 	ascending bool,
-	limit uint,
+	limit uint64,
 	lastOrderedValue *float64,
 	lastID *uint64,
 ) ([]TorrentMetadata, error) {
