+++ persistence/sqlite3_test.go
@@ -127,7 +127,7 @@ func Test_sqlite3Database_QueryTorrents(t *testing.T) {
 		epoch            int64
 		orderBy          OrderingCriteria
 		ascending        bool
-		limit            uint
+		limit            uint64
 		lastOrderedValue *float64
 		lastID           *uint64
 		want             []TorrentMetadata
